requirements.txt
rich>=13.7.0
colorama>=0.4.6
requests>=2.31.0
reportlab>=4.1.0
python-docx>=1.1.0
psutil>=5.9.8